+++ server/src/main/java/org/opensearch/indices/recovery/RecoveryState.java
@@ -50,6 +50,7 @@ import org.opensearch.common.xcontent.XContentFactory;
 import org.opensearch.index.shard.IndexShard;
 import org.opensearch.index.shard.ShardId;
 import org.opensearch.index.store.StoreStats;
+import org.opensearch.indices.replication.common.Timer;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -369,77 +370,6 @@ public class RecoveryState implements ToXContentFragment, Writeable {
         static final String SOURCE_THROTTLE_TIME_IN_MILLIS = "source_throttle_time_in_millis";
         static final String TARGET_THROTTLE_TIME = "target_throttle_time";
         static final String TARGET_THROTTLE_TIME_IN_MILLIS = "target_throttle_time_in_millis";
-    }
-
-    public static class Timer implements Writeable {
-        private long startTime = 0;
-        private long startNanoTime = 0;
-        private long time = -1;
-        private long stopTime = 0;
-
-        public Timer() {}
-
-        public Timer(StreamInput in) throws IOException {
-            startTime = in.readVLong();
-            startNanoTime = in.readVLong();
-            stopTime = in.readVLong();
-            time = in.readVLong();
-        }
-
-        @Override
-        public synchronized void writeTo(StreamOutput out) throws IOException {
-            out.writeVLong(startTime);
-            out.writeVLong(startNanoTime);
-            out.writeVLong(stopTime);
-            // write a snapshot of current time, which is not per se the time field
-            out.writeVLong(time());
-        }
-
-        public synchronized void start() {
-            assert startTime == 0 : "already started";
-            startTime = System.currentTimeMillis();
-            startNanoTime = System.nanoTime();
-        }
-
-        /** Returns start time in millis */
-        public synchronized long startTime() {
-            return startTime;
-        }
-
-        /** Returns elapsed time in millis, or 0 if timer was not started */
-        public synchronized long time() {
-            if (startNanoTime == 0) {
-                return 0;
-            }
-            if (time >= 0) {
-                return time;
-            }
-            return Math.max(0, TimeValue.nsecToMSec(System.nanoTime() - startNanoTime));
-        }
-
-        /** Returns stop time in millis */
-        public synchronized long stopTime() {
-            return stopTime;
-        }
-
-        public synchronized void stop() {
-            assert stopTime == 0 : "already stopped";
-            stopTime = Math.max(System.currentTimeMillis(), startTime);
-            time = TimeValue.nsecToMSec(System.nanoTime() - startNanoTime);
-            assert time >= 0;
-        }
-
-        public synchronized void reset() {
-            startTime = 0;
-            startNanoTime = 0;
-            time = -1;
-            stopTime = 0;
-        }
-
-        // for tests
-        public long getStartNanoTime() {
-            return startNanoTime;
-        }
     }
 
     public static class VerifyIndex extends Timer implements ToXContentFragment, Writeable {
