+++ server/src/test/java/org/opensearch/indices/recovery/RecoveryTargetTests.java
@@ -44,9 +44,9 @@ import org.opensearch.index.shard.ShardId;
 import org.opensearch.indices.recovery.RecoveryState.FileDetail;
 import org.opensearch.indices.recovery.RecoveryState.Index;
 import org.opensearch.indices.recovery.RecoveryState.Stage;
-import org.opensearch.indices.recovery.RecoveryState.Timer;
 import org.opensearch.indices.recovery.RecoveryState.Translog;
 import org.opensearch.indices.recovery.RecoveryState.VerifyIndex;
+import org.opensearch.indices.replication.common.Timer;
 import org.opensearch.test.OpenSearchTestCase;
 
 import java.io.IOException;
